Maintnance of MainWindow code, move to folder files, add to main first line for MDI, add header with string declaration
