+++ GlobalNames.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <QString>
+#include <QTranslator>
+
+namespace GLOBAL_STRINGS
+{
+    namespace PROGRAM_INFO {
+        QString PROGRAM_NAME(QObject::tr("spacePaint"));
+        QString ORGANIZATION_NAME(QObject::tr("DB&DG"));
+        QString PROGRAM_VERSION(QObject::tr("0.1"));
+    }
+}
