+++ main.cpp
@@ -1,11 +1,22 @@
-#include "mainwindow.h"
+#include "MainWindow/MainWindow.h"
 #include <QApplication>
+
+#include "GlobalNames.h"
 
 int main(int argc, char *argv[])
 {
-    QApplication a(argc, argv);
-    MainWindow w;
-    w.show();
+    //Q_INIT_RESOURCE(spacePaint);
 
-    return a.exec();
+    QApplication Application(argc, argv);
+
+    QCoreApplication::setApplicationName(GLOBAL_STRINGS::PROGRAM_INFO::PROGRAM_NAME);
+    QCoreApplication::setOrganizationDomain(GLOBAL_STRINGS::PROGRAM_INFO::ORGANIZATION_NAME);
+    QCoreApplication::setApplicationName(GLOBAL_STRINGS::PROGRAM_INFO::PROGRAM_VERSION);
+
+
+
+    MainWindow ApplicationWindow;
+    ApplicationWindow.show();
+
+    return Application.exec();
 }
